Move generic utility function
