+++ src/network/GossipManager.ts
@@ -1,9 +1,8 @@
 import {ArraySet} from "array-map-set";
 import {EventEmitter} from "events";
 import {hash} from "../crypto/crypto";
-import {bin2Hex} from "../utils/utils";
+import {bin2Hex, compareUint8Array} from "../utils/utils";
 import {COMMANDS, COMMANDS_INVERSE, GOSSIP_COMMANDS, ICommandsKeys} from "./commands";
-import {compareUint8Array} from "./Network";
 import {TcpManager} from "./TcpManager";
 import {UdpManager} from "./UdpManager";
 import {composeMessage, noopResponseCallback} from "./utils";
